+++ src/main/java/org/plumelib/javadoc/RequireJavadoc.java
@@ -1,326 +1,442 @@
 package org.plumelib.javadoc;
 
-import com.sun.javadoc.AnnotationDesc;
-import com.sun.javadoc.ClassDoc;
-import com.sun.javadoc.ConstructorDoc;
-import com.sun.javadoc.Doc;
-import com.sun.javadoc.DocErrorReporter;
-import com.sun.javadoc.FieldDoc;
-import com.sun.javadoc.MethodDoc;
-import com.sun.javadoc.PackageDoc;
-import com.sun.javadoc.RootDoc;
-import com.sun.javadoc.SourcePosition;
-import com.sun.tools.doclets.standard.Standard;
+import static com.github.javaparser.utils.PositionUtils.sortByBeginPosition;
+
+import com.github.javaparser.ParseProblemException;
+import com.github.javaparser.Position;
+import com.github.javaparser.Range;
+import com.github.javaparser.StaticJavaParser;
+import com.github.javaparser.ast.CompilationUnit;
+import com.github.javaparser.ast.Node;
+import com.github.javaparser.ast.PackageDeclaration;
+import com.github.javaparser.ast.body.ClassOrInterfaceDeclaration;
+import com.github.javaparser.ast.body.ConstructorDeclaration;
+import com.github.javaparser.ast.body.FieldDeclaration;
+import com.github.javaparser.ast.body.MethodDeclaration;
+import com.github.javaparser.ast.body.VariableDeclarator;
+import com.github.javaparser.ast.comments.Comment;
+import com.github.javaparser.ast.expr.AnnotationExpr;
+import com.github.javaparser.ast.nodeTypes.NodeWithJavadoc;
+import com.github.javaparser.ast.visitor.VoidVisitorAdapter;
 import java.io.File;
+import java.io.IOException;
+import java.nio.file.FileVisitResult;
+import java.nio.file.FileVisitor;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.nio.file.SimpleFileVisitor;
+import java.nio.file.attribute.BasicFileAttributes;
 import java.util.ArrayList;
-import java.util.HashSet;
+import java.util.Comparator;
+import java.util.LinkedList;
 import java.util.List;
-import java.util.Set;
-import java.util.TreeSet;
+import java.util.Optional;
 import java.util.regex.Pattern;
-import org.checkerframework.checker.formatter.qual.Format;
 import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
-import org.checkerframework.checker.nullness.qual.NonNull;
-import org.checkerframework.checker.regex.RegexUtil;
-import org.checkerframework.common.value.qual.MinLen;
+import org.plumelib.options.Option;
+import org.plumelib.options.Options;
 
 /**
- * A Javadoc doclet that issues an error for any package, class, method, or field that lacks a
- * Javadoc comment.
+ * A program that issues an error for any class, constructor, method, or field that lacks a Javadoc
+ * comment. Does not issue a warning for methods annotated with {@code @Override}.
  */
-// This doclet extends the standard doclet so that this doclet accepts the standard command-line
-// arguments (even though it ignores them), which makes this doclet easier to use in an existing
-// build or with a build system.
-public class RequireJavadoc extends Standard {
+public class RequireJavadoc {
+
+  /** All the errors this program will report. */
+  List<String> errors = new ArrayList<>();
+
+  /**
+   * Matches name of file or directory class/constructor/method/field, or full package name, where
+   * no problems should be reported.
+   */
+  @Option("Don't check files or directories whose pathname matches the regex")
+  public @MonotonicNonNull Pattern exclude = null;
+
+  // TODO: It would be nice to support matching fully-qualified class names, but matching
+  // packages will have to do for now.
+  /**
+   * Matches simple name of class/constructor/method/field, or full package name, where no problems
+   * should be reported.
+   */
+  @Option("Don't report problems in Java elements whose name matches the regex")
+  public @MonotonicNonNull Pattern dont_require = null;
+
+  /**
+   * If true, print filenames relative to working directory. Setting this only has an effect if the
+   * command-line arguments were absolute pathnames, or no command-line arguments were supplied.
+   */
+  @Option("Report relative rather than absolute filenames")
+  public boolean relative = false;
 
   /** If true, output debug information. */
-  private static boolean verbose = false;
+  @Option("Print diagnostic information")
+  public boolean verbose = false;
 
-  /** All the errors this doclet will report. */
-  TreeSet<String> errors = new TreeSet<>();
-
-  /**
-   * All packages that have been seen and checked so far. This helps prevent issuing multiple
-   * warnings about a given package.
-   */
-  Set<PackageDoc> packages = new HashSet<>();
-
-  /** If true, print filenames relative to working directory. If false, print absolute paths. */
-  static boolean relativePaths = false;
-
-  /** If non-null, matches classes where no problems should be reported. */
-  static @MonotonicNonNull Pattern skip = null;
+  /** The Java files to be checked. */
+  List<Path> javaFiles = new ArrayList<Path>();
 
   /** The current working directory, for making relative pathnames. */
-  static Path currentPath = Paths.get("").toAbsolutePath();
+  Path workingDirRelative = Paths.get("");
 
-  private static final @Format({}) String USAGE =
-      "Provided by RequireJavadoc doclet:%n"
-          + "-skip <regex>      Don't report problems in classes that match the given regex%n"
-          + "-relative          Report relative rather than absolute filenames%n"
-          + "verbose            Print diagnostic information%n"
-          + "See the documentation for more details.%n";
+  /** The current working directory, for making relative pathnames. */
+  Path workingDirAbsolute = Paths.get("").toAbsolutePath();
+
+  /**
+   * The main entry point for the require-javadoc program.
+   *
+   * @param args the command-line arguments; see the README file.
+   */
+  public static void main(String[] args) {
+    RequireJavadoc rj = new RequireJavadoc();
+    Options options =
+        new Options(
+            "java org.plumelib.javadoc.RequireJavadoc [options] [directory-or-file ...]", rj);
+    String[] remainingArgs = options.parse(true, args);
+
+    rj.setJavaFiles(remainingArgs);
+
+    for (Path javaFile : rj.javaFiles) {
+      try {
+        CompilationUnit cu = StaticJavaParser.parse(javaFile);
+        Optional<PackageDeclaration> opd = cu.getPackageDeclaration();
+        if (opd.isPresent()) {
+          String packageName = opd.get().getName().asString();
+          if (rj.shouldNotRequire(packageName)) {
+            continue;
+          }
+        }
+        RequireJavadocVisitor visitor = rj.new RequireJavadocVisitor(javaFile);
+        visitor.visit(cu, null);
+      } catch (IOException e) {
+        System.out.println("Problem while reading " + javaFile + ": " + e.getMessage());
+        System.exit(2);
+      } catch (ParseProblemException e) {
+        System.out.println("Problem while parsing " + javaFile + ": " + e.getMessage());
+        System.exit(2);
+      }
+    }
+    for (String error : rj.errors) {
+      System.out.println(error);
+    }
+    System.exit(rj.errors.isEmpty() ? 0 : 1);
+  }
 
   /** Creates a new RequireJavadoc instance. */
-  RequireJavadoc() {}
+  private RequireJavadoc() {}
 
   /**
-   * The entry point to this doclet. Requires documentation of all Java packages, classes, methods,
-   * and fields.
+   * Set the Java files to be processed from the command-line arguments.
    *
-   * <p>{@inheritDoc}
+   * @param args the directories and files listed on the command line
    */
-  public static boolean start(RootDoc root) {
-    RequireJavadoc doclet = new RequireJavadoc();
-    if (verbose) {
-      System.err.printf("start(): root=%s%n", root);
-      for (ClassDoc cd : root.classes()) {
-        System.err.printf("  %s%n", cd);
+  @SuppressWarnings("lock:methodref.receiver.invalid") // no locking here
+  private void setJavaFiles(String[] args) {
+    if (args.length == 0) {
+      args = new String[] {workingDirAbsolute.toString()};
+    }
+
+    FileVisitor<Path> walker = new JavaFilesVisitor();
+
+    for (String arg : args) {
+      if (shouldExclude(arg)) {
+        continue;
+      }
+      Path p = Paths.get(arg);
+      File f = p.toFile();
+      if (!f.exists()) {
+        System.out.println("File not found: " + f);
+        System.exit(2);
+      }
+      if (f.isDirectory()) {
+        try {
+          Files.walkFileTree(p, walker);
+        } catch (IOException e) {
+          System.out.println("Problem while reading " + f + ": " + e.getMessage());
+          System.exit(2);
+        }
+      } else {
+        javaFiles.add(Paths.get(arg));
       }
     }
-    for (ClassDoc cd : root.classes()) {
-      doclet.processClass(cd);
-    }
 
-    for (String error : doclet.errors) {
-      System.err.println(error);
-    }
-
-    return doclet.errors.isEmpty();
+    javaFiles.sort(Comparator.comparing(Object::toString));
   }
 
-  /**
-   * Require documentation of the given class and its package.
-   *
-   * @param cd the class to require documentation of
-   */
-  private void processClass(ClassDoc cd) {
-    if (verbose) {
-      System.err.printf("processClass(%s)%n", cd);
-    }
-    SourcePosition classPosition = cd.position();
-    if (skip != null
-        && (skip.matcher(cd.name()).find()
-            || skip.matcher(cd.qualifiedName()).find()
-            || (classPosition != null && skip.matcher(classPosition.file().toString()).find()))) {
-      if (verbose) {
-        System.err.printf("processClass: skipped %s%n", cd);
-      }
-      return;
-    }
-    requireCommentText(cd);
-    PackageDoc pd = cd.containingPackage();
-    // TODO: This does not work: it does not require documentation for packages.
-    if (packages.add(pd)) {
-      requireCommentText(pd);
-    }
-    for (ConstructorDoc constructorDoc : cd.constructors()) {
-      // Don't require documentation of synthetic constructors, which don't appear in the source
-      // code.
-      if (!constructorDoc.isSynthetic() && !isSyntheticForNestedConstructor(constructorDoc)) {
-        requireCommentText(constructorDoc);
-      }
-    }
-    for (FieldDoc ed : cd.enumConstants()) {
-      requireCommentText(ed);
-    }
-    if (verbose) {
-      System.err.printf("Class %s has %s fields%n", cd, cd.fields().length);
-    }
-    for (FieldDoc fd : cd.fields()) {
-      // Don't require documentation for `long serialVersionUID`.
-      if (!(fd.name().equals("serialVersionUID") && fd.type().toString().equals("long"))) {
-        requireCommentText(fd);
-      }
-    }
-    for (ClassDoc icd : cd.innerClasses()) {
-      requireCommentText(icd);
-    }
-    if (verbose) {
-      System.err.printf("Class %s has %s methods%n", cd, cd.methods().length);
-    }
-    for (MethodDoc md : cd.methods()) {
-      // Don't require documentation of overriding or synthetic methods
-      if (!isOverride(md) && !md.isSynthetic() && !isEnumValuesOrValueOf(md)) {
-        requireCommentText(md);
-      } else {
-        if (verbose) {
-          System.err.printf(
-              "Method %s not checked: override=%s synthetic=%s enum=%s%n",
-              md, isOverride(md), md.isSynthetic(), isEnumValuesOrValueOf(md));
+  /** Collects files into the {@link #javaFiles} variable. */
+  class JavaFilesVisitor extends SimpleFileVisitor<Path> {
+
+    @Override
+    public FileVisitResult visitFile(Path file, BasicFileAttributes attr) {
+      if (!shouldExclude(file)) {
+        if (attr.isRegularFile() && file.toString().endsWith(".java")) {
+          javaFiles.add(file);
         }
       }
+      return FileVisitResult.CONTINUE;
+    }
+
+    @Override
+    public FileVisitResult preVisitDirectory(Path dir, BasicFileAttributes attr) {
+      if (shouldExclude(dir)) {
+        return FileVisitResult.SKIP_SUBTREE;
+      }
+      return FileVisitResult.CONTINUE;
+    }
+
+    @Override
+    public FileVisitResult postVisitDirectory(Path dir, IOException exc) {
+      if (exc != null) {
+        System.out.println("Problem visiting " + dir + ": " + exc.getMessage());
+        System.exit(2);
+      }
+      return FileVisitResult.CONTINUE;
+    }
+
+    @Override
+    public FileVisitResult visitFileFailed(Path file, IOException exc) {
+      if (exc != null) {
+        System.out.println("Problem visiting " + file + ": " + exc.getMessage());
+        System.exit(2);
+      }
+      return FileVisitResult.CONTINUE;
     }
   }
 
   /**
-   * Return true if this method is annotated with {@code @Override}.
+   * Return true if the given Java element should not be checked.
    *
-   * @param md the method to check for an {@code @Override} annotation
-   * @return true if this method is annotated with {@code @Override}
+   * @param name the name of a Java element
+   * @return true if no warnings should be issued about the element
    */
-  private boolean isOverride(MethodDoc md) {
-    for (AnnotationDesc anno : md.annotations()) {
-      // TODO: A String comparison is a bit gross.
-      if (anno.toString().equals("@java.lang.Override")) {
+  // `name` is a simple name for elements other than packages.
+  boolean shouldNotRequire(String name) {
+    boolean result = dont_require != null && dont_require.matcher(name).find();
+    if (verbose) {
+      System.out.printf("shouldNotRequire(%s) => %s%n", name, result);
+    }
+    return result;
+  }
+
+  /**
+   * Return true if the given file or directory should be skipped.
+   *
+   * @param fileName the name of a Java file or directory
+   * @return true if the file or directory should be skipped
+   */
+  boolean shouldExclude(String fileName) {
+    boolean result = exclude != null && exclude.matcher(fileName).find();
+    if (verbose) {
+      System.out.printf("shouldExclude(%s) => %s%n", fileName, result);
+    }
+    return result;
+  }
+
+  /**
+   * Return true if the given file or directory should be skipped.
+   *
+   * @param path a Java file or directory
+   * @return true if the file or directory should be skipped
+   */
+  boolean shouldExclude(Path path) {
+    return shouldExclude(path.toString());
+  }
+
+  /** Visits an AST and collects warnings about missing Javadoc. */
+  private class RequireJavadocVisitor extends VoidVisitorAdapter<Void> {
+
+    /** The file being visited. Used for constructing error messages. */
+    Path filename;
+
+    /**
+     * Create a new RequireJavadocVisitor.
+     *
+     * @param filename the file being visited; used for diagnostic messages
+     */
+    RequireJavadocVisitor(Path filename) {
+      this.filename = filename;
+    }
+
+    /**
+     * Return a string stating that documentation is missing on the given construct.
+     *
+     * @param node a Java language construct (class, constructor, method, field)
+     * @param simpleName the construct's simple name, used in diagnostic messages
+     * @return an error message for the given construct
+     */
+    private String errorString(Node node, String simpleName) {
+      Optional<Range> range = node.getRange();
+      if (range.isPresent()) {
+        Position begin = range.get().begin;
+        Path path =
+            (relative
+                ? (filename.isAbsolute() ? workingDirAbsolute : workingDirRelative)
+                    .relativize(filename)
+                : filename);
+        return String.format(
+            "%s:%d:%d: missing documentation for %s", path, begin.line, begin.column, simpleName);
+      } else {
+        return "missing documentation for " + simpleName;
+      }
+    }
+
+    @Override
+    public void visit(ClassOrInterfaceDeclaration cd, Void ignore) {
+      String name = cd.getNameAsString();
+      if (shouldNotRequire(name)) {
+        return;
+      }
+      super.visit(cd, ignore);
+      if (verbose) {
+        System.out.printf("Visiting %s%n", cd.getName());
+      }
+      if (!hasJavadocComment(cd)) {
+        errors.add(errorString(cd, name));
+      }
+    }
+
+    @Override
+    public void visit(ConstructorDeclaration cd, Void ignore) {
+      String name = cd.getNameAsString();
+      if (shouldNotRequire(name)) {
+        return;
+      }
+      super.visit(cd, ignore);
+      if (verbose) {
+        System.out.printf("Visiting %s%n", cd.getName());
+      }
+      if (!hasJavadocComment(cd)) {
+        errors.add(errorString(cd, name));
+      }
+    }
+
+    @Override
+    public void visit(MethodDeclaration md, Void ignore) {
+      String name = md.getNameAsString();
+      if (shouldNotRequire(name)) {
+        return;
+      }
+      super.visit(md, ignore);
+      if (verbose) {
+        System.out.printf("Visiting %s%n", md.getName());
+      }
+      if (isOverride(md)) {
+        return;
+      }
+      if (!hasJavadocComment(md)) {
+        errors.add(errorString(md, name));
+      }
+    }
+
+    @Override
+    public void visit(FieldDeclaration fd, Void ignore) {
+      boolean shouldSuperVisit = false;
+      if (verbose) {
+        System.out.printf("Visiting %s%n", fd.getVariables().get(0).getName());
+      }
+      boolean hasJavadocComment = hasJavadocComment(fd);
+      for (VariableDeclarator vd : fd.getVariables()) {
+        String name = vd.getNameAsString();
+        if (shouldNotRequire(name)) {
+          continue;
+        }
+        shouldSuperVisit = true;
+        // TODO: Also check the type of the serialVersionUID variable.
+        if (name.equals("serialVersionUID")) {
+          continue;
+        }
+        if (!hasJavadocComment) {
+          errors.add(errorString(vd, name));
+        }
+      }
+      if (shouldSuperVisit) {
+        super.visit(fd, ignore);
+      }
+    }
+
+    /**
+     * Return true if this method is annotated with {@code @Override}.
+     *
+     * @param md the method to check for an {@code @Override} annotation
+     * @return true if this method is annotated with {@code @Override}
+     */
+    private boolean isOverride(MethodDeclaration md) {
+      for (AnnotationExpr anno : md.getAnnotations()) {
+        if (anno.toString().equals("@Override") || anno.toString().equals("@java.lang.Override")) {
+          return true;
+        }
+      }
+      return false;
+    }
+  }
+
+  /**
+   * Return true if this node has a Javadoc comment.
+   *
+   * @param n the node to check for a Javadoc comment
+   * @return true if this node has a Javadoc comment
+   */
+  boolean hasJavadocComment(Node n) {
+    if (n instanceof NodeWithJavadoc && ((NodeWithJavadoc<?>) n).hasJavaDocComment()) {
+      return true;
+    }
+    List<Comment> orphans = new ArrayList<>();
+    getOrphanCommentsBeforeThisChildNode(n, orphans);
+    for (Comment orphan : orphans) {
+      if (orphan.isJavadocComment()) {
         return true;
       }
     }
     return false;
   }
 
+  // This method comes from Randoop's `Minimize.java` file.  Randoop had adapted it from JavaParser.
   /**
-   * Return true if this method is a synthetic default constructor. The <a
-   * href="https://docs.oracle.com/javase/8/docs/jdk/api/javadoc/doclet/com/sun/javadoc/MemberDoc.html">isSynthetic</a>
-   * documentation says "Returns true if this member was synthesized by the compiler.", but it
-   * returns false for a synthesized default constructor for a nested class. This handles that case.
+   * This is stolen from JavaParser's PrettyPrintVisitor.printOrphanCommentsBeforeThisChildNode,
+   * with light modifications.
    *
-   * @param cd the constructor to check
-   * @return true if {@code cd} is a synthetic constructor
-   */
-  private boolean isSyntheticForNestedConstructor(ConstructorDoc cd) {
-    SourcePosition constructorPosition = cd.position();
-    if (constructorPosition == null) {
-      return false;
-    }
-    @SuppressWarnings("nullness:assignment.type.incompatible") // constructor has a containing class
-    @NonNull ClassDoc containingClass = cd.containingClass();
-    SourcePosition classPosition = containingClass.position();
-    assert classPosition != null : "@AssumeAssertion(nullness): a class has a position";
-    return classPosition.file().equals(constructorPosition.file())
-        && classPosition.line() == constructorPosition.line();
-  }
-
-  /**
-   * Return true if the method is {@code values} or {@code valueOf} for an enum.
-   *
-   * @param md the MethodDoc to test
-   * @return true if the method is {@code values} or {@code valueOf} for an enum
+   * @param node the node whose orphan comments to collect
+   * @param result the list to add orphan comments to. Is side-effected by this method. The
+   *     implementation uses this to minimize the diffs against upstream.
    */
   @SuppressWarnings({
-    "index:array.access.unsafe.high.constant", // MethodDoc.parameters() needs @Pure annotation
-    "nullness:dereference.of.nullable" // md.containingClass() is non-null for a MethodDoc
-    // (MethodDoc needs annotation)
+    "JdkObsolete", // for LinkedList
+    "interning:not.interned", // element of a list
+    "ReferenceEquality",
   })
-  private boolean isEnumValuesOrValueOf(MethodDoc md) {
-    return md.containingClass().isEnum()
-        && ((md.name().equals("values") && md.parameters().length == 0)
-            || (md.name().equals("valueOf")
-                && md.parameters().length == 1
-                // String comparison is gross, but this is a Javadoc type rather than a
-                // reflection type.
-                && md.parameters()[0].type().qualifiedTypeName().equals("java.lang.String")));
-  }
-
-  /**
-   * Require a documentation string for the given Java element. If there is no documentation, this
-   * doclet will print a warning message and will return an error status.
-   *
-   * @param d any Java element
-   */
-  private void requireCommentText(Doc d) {
-    if (skip != null && skip.matcher(d.name()).find()) {
-      if (verbose) {
-        System.err.printf("requireCommentText(%s) SKIPPED%n", d);
-      }
+  private static void getOrphanCommentsBeforeThisChildNode(final Node node, List<Comment> result) {
+    if (node instanceof Comment) {
       return;
     }
-    if (verbose) {
-      System.err.printf("requireCommentText(%s)%n", d);
-    }
-    String text = d.getRawCommentText();
-    if (text.isEmpty()) {
-      if (!d.name().equals("")) { // don't warn about the unnamed package
-        errors.add(errorString(d));
-      }
-    }
-  }
 
-  /**
-   * Return a string stating that documentation is missing on the given construct.
-   *
-   * @param d a Java language construct (package, class, constructor, method, field)
-   * @return an error message for the given construct
-   */
-  private String errorString(Doc d) {
-    SourcePosition position = d.position();
-    if (position == null) {
-      return "missing documentation for " + d.name();
+    Node parent = node.getParentNode().orElse(null);
+    if (parent == null) {
+      return;
     }
-    File file = position.file();
-    // Making the pathname relative shortens it, but can confuse tools.
-    Path path = (relativePaths ? currentPath.relativize(file.toPath()) : file.toPath());
-    return String.format("%s:%d: missing documentation for %s", path, position.line(), d.name());
-  }
-
-  /**
-   * Given a command-line option of this doclet, returns the number of arguments you must specify on
-   * the command line for the given option. Returns 0 if the argument is not recognized. This method
-   * is automatically invoked by Javadoc.
-   *
-   * @param option the command-line option
-   * @return the number of command-line arguments needed when using the option
-   * @see <a
-   *     href="https://docs.oracle.com/javase/8/docs/technotes/guides/javadoc/doclet/overview.html">Doclet
-   *     overview</a>
-   */
-  public static int optionLength(String option) {
-    switch (option) {
-      case "-relative":
-        return 1;
-      case "-skip":
-        return 2;
-      case "-verbose":
-        return 1;
-      default:
-        return Standard.optionLength(option);
+    List<Node> everything = new LinkedList<>(parent.getChildNodes());
+    sortByBeginPosition(everything);
+    int positionOfTheChild = -1;
+    for (int i = 0; i < everything.size(); i++) {
+      if (everything.get(i) == node) positionOfTheChild = i;
     }
-  }
-
-  /**
-   * Tests the validity of command-line arguments passed to this doclet. Returns true if the option
-   * usage is valid, and false otherwise. This method is automatically invoked by Javadoc.
-   *
-   * <p>Also sets fields from the command-line arguments.
-   *
-   * @param options the command-line options to be checked: an array of 1- or 2-element arrays,
-   *     where the length depends on {@link #optionLength} applied to the first element
-   * @param reporter where to report errors
-   * @return true iff the command-line options are valid
-   * @see <a
-   *     href="https://docs.oracle.com/javase/8/docs/technotes/guides/javadoc/doclet/overview.html">Doclet
-   *     overview</a>
-   */
-  public static boolean validOptions(String[] @MinLen(1) [] options, DocErrorReporter reporter) {
-    List<String[]> remaining = new ArrayList<>();
-    for (int oi = 0; oi < options.length; oi++) {
-      String[] os = options[oi];
-      String opt = os[0].toLowerCase();
-      switch (opt) {
-        case "-relative":
-          relativePaths = true;
-          break;
-        case "-skip":
-          assert os.length == 2 : "@AssumeAssertion(value): dependent: optionLength(\"-skip\")==2";
-          if (!RegexUtil.isRegex(os[1])) {
-            System.err.printf("Error parsing regex %s %s%n", os[1], RegexUtil.regexError(os[1]));
-            System.exit(2);
-          }
-          skip = Pattern.compile(os[1]);
-          break;
-        case "-verbose":
-          verbose = true;
-          break;
-        case "-help":
-          System.out.printf(USAGE);
-          return false;
-        default:
-          remaining.add(os);
-          break;
-      }
+    if (positionOfTheChild == -1) {
+      throw new AssertionError("I am not a child of my parent.");
     }
-    return Standard.validOptions(remaining.toArray(new String[0][]), reporter);
+    int positionOfPreviousChild = -1;
+    for (int i = positionOfTheChild - 1; i >= 0 && positionOfPreviousChild == -1; i--) {
+      if (!(everything.get(i) instanceof Comment)) positionOfPreviousChild = i;
+    }
+    for (int i = positionOfPreviousChild + 1; i < positionOfTheChild; i++) {
+      Node nodeToPrint = everything.get(i);
+      if (!(nodeToPrint instanceof Comment))
+        throw new RuntimeException(
+            "Expected comment, instead "
+                + nodeToPrint.getClass()
+                + ". Position of previous child: "
+                + positionOfPreviousChild
+                + ", position of child "
+                + positionOfTheChild);
+      result.add((Comment) nodeToPrint);
+    }
   }
 }